sentry: Extract functions used by channel into a util file.
This is to make tests for channel work.

Channel mocks `$` to be an object, so it cannot be called as an
function in sentry.ts if it is imported in channel.ts.

So, to avoid it, we extract the functions which channel.ts uses
and fortunately they don't use `$`, so it works out.
