+++ tools/test-js-with-node
@@ -192,6 +192,7 @@ EXEMPT_FILES = make_set(
         "web/src/search.js",
         "web/src/sent_messages.ts",
         "web/src/sentry.ts",
+        "web/src/sentry_util.ts",
         "web/src/server_events.js",
         "web/src/settings.js",
         "web/src/settings_account.js",
