+++ web/src/sentry_util.ts
@@ -1,0 +1,20 @@
+export function normalize_path(path: string, is_portico = false): string {
+    if (path === undefined) {
+        return "unknown";
+    }
+    path = path
+        .replace(/\/\d+(\/|$)/, "/*$1")
+        .replace(
+            /^\/(join|reactivate|new|accounts\/do_confirm|accounts\/confirm_new_email)\/[^/]+(\/?)$/,
+            "$1/*$2",
+        );
+    if (is_portico) {
+        return "portico: " + path;
+    }
+    return path;
+}
+
+export function shouldCreateSpanForRequest(url: string): boolean {
+    const parsed = new URL(url, window.location.href);
+    return parsed.pathname !== "/json/events";
+}
